FFM preprocessing done

+++ benchmark_process_private.py
@@ -9,6 +9,8 @@ from dask.diagnostics import ProgressBar
 from dask.distributed import Client,LocalCluster
 from sklearn.feature_extraction import FeatureHasher
 from sklearn.compose import ColumnTransformer
+from KFT.benchmarks.utils import core_data_extract_df
+
 PATH = './raw_data_hm/'
 np.random.seed(1337)
 
@@ -126,14 +128,53 @@ if __name__ == '__main__':
         df= pd.concat([df,to_be_hashed],axis=1)
         df = dd.from_pandas(df,npartitions=64)
         df.to_parquet('./benchmark_data_private_hashed/')
-    df = dd.read_parquet('./benchmark_data_private_hashed/')
-    Y = df['total_sales'].compute()
-    X = df.drop('total_sales',axis=1)
-    print(X)
-    s = StandardScaler()
-    X = s.fit_transform(X).compute()
-    df = pd.concat([X, Y], axis=1)
-    df = dd.from_pandas(df,npartitions=64)
-    df.to_parquet('./benchmark_data_private_hashed_scaled/')
+    if not os.path.exists('benchmark_data_private_hashed'):
+        df = dd.read_parquet('./benchmark_data_private_hashed/')
+        Y = df['total_sales'].compute()
+        X = df.drop('total_sales',axis=1)
+        print(X)
+        s = StandardScaler()
+        X = s.fit_transform(X).compute()
+        df = pd.concat([X, Y], axis=1)
+        df = dd.from_pandas(df,npartitions=64)
+        df.to_parquet('./benchmark_data_private_hashed_scaled/')
+
+
+    if not os.path.exists('benchmark_data_private_FFM'):
+        df = dd.read_parquet('./benchmark_data_lgbm/')  # .compute()
+        categorical_columns = [
+            'location_id',
+            'city_id',
+            'corporate_brand_id',
+            'graphical_appearance_id',
+            'colour_id',
+            'enterprise_size_id',
+            'department_id',
+            'product_season_id',
+            'product_type_id',
+            'product_group_no',
+            'product_id',
+            'article_id',
+        ]
+        n_cat = 100
+        df = df.reset_index(drop=True)
+        Y = df['total_sales'].compute()
+        X = df.drop('total_sales',axis=1).compute()
+        df_num = X.select_dtypes(include=[np.float]).columns
+        for col in X.columns:
+            print(X[col].unique())
+            X[col] = X[col].fillna(X[col].mean())
+        for el in df_num:
+            X[el] =pd.cut(X[el],n_cat,labels=False)
+        print(X.columns[-1])
+        X = core_data_extract_df(X)
+        X = X.reset_index(drop=True)
+        Y=Y.reset_index(drop=True)
+        df = pd.concat([X,Y],axis=1)
+        df = dd.from_pandas(df, npartitions=64)
+        df.to_parquet('./benchmark_data_private_FFM/')
+    df = dd.read_parquet('./benchmark_data_private_FFM/').compute()
+    print(df.nunique())
+
